test(tui): split PTY harness support
